+++ fixtures/parser_tests.js
@@ -7,6 +7,7 @@
 
 const fixtures = __dirname;
 
+// eslint-disable-next-line no-undef
 function parserTests(parse: (file: string) => BabylonParserResult) {
   describe('File parsing without throwing', () => {
     it('Should not throw', () => {